build: functions as multi-entry library
